map API implementation
Hi,
I created the geo dir , with GeoConfig , GeoService, GeoController. update build.gradle, the google map api key in application.yml .

We can now get position in LagLng from an calling address .
There are also two way to get distance , with straight line or a driving path . The straight line return distance in meter (type long) . The drivepath is a record dto, with time , distance , and a list of positions ( turns on the path).
there is also a simple method returning the dto from the two input address. It could be useful for easier Price calculation .